fix: Room Accommodation 클래스 및 클래스 다이어그램 수정
